moved label-count panel to sep plugin
