+++ plugins/data_quality_panel
@@ -1,0 +1,1 @@
+/Users/ritchie/Projects/voxel51/data_quality_panel
